services/source_fetcher/requirements.txt
sqlalchemy
aiomysql
pika
feedparser
aiohttp
python-dotenv
python-json-logger